review followups: spool ceil, pointermove + marker re-render, fetch abort
- fitting/engine.py: cycles_to_max ceil instead of int truncate (off-by-one on time-to-spool-max).
- StarMap.tsx: pointermove now updates React state only when hovered system ID changes — quadtree.find returns a fresh object each call so prior code re-rendered the whole tree at 60Hz during cursor motion. Renderer-level highlight calls still fire every move.
- StarMap.tsx: introduce vpVersion state bumped on viewport moved/zoomed so character-location markers re-position during pan/zoom even when no system is selected (previously only updated as a side-effect of panelPos for selected systems).
- useMapData.ts: AbortController on the K-space and W-space JSON fetches so a rapid space toggle doesn't leave stale requests in flight.
- useKillHeatmap.ts: Uint16Array → Uint32Array (a busy null hub during a war can exceed 65535 kills per bucket and silently wrap).
- useBookmarks.ts: temp ID via monotonic counter ref instead of -Date.now() (collision on rapid double-clicks within the same millisecond would replace both adds with the first server response).

Spool-up SDE units verified: attr 2734 damageMultiplierBonusMax for Light/Heavy/Supratidal Entropic Disintegrator I = 2.125, producing the in-game 3.125× volley ceiling via 1 + 2.125. Stored as a fraction; existing volley * (1 + spool_max) formula is correct.

+++ frontend/src/components/StarMap/useKillHeatmap.ts
@@ -20,15 +20,15 @@ export interface KillHeatmapData {
   buckets: string[];
   max_value: number;
   /** Per-system dense lookup array. Length = buckets.length. */
-  lookup: Map<number, Uint16Array>;
+  lookup: Map<number, Uint32Array>;
 }
 
 function materialize(raw: KillHeatmapResponseRaw): KillHeatmapData {
   const numBuckets = raw.buckets.length;
-  const lookup = new Map<number, Uint16Array>();
+  const lookup = new Map<number, Uint32Array>();
   if (raw.sparse) {
     for (const [sid, pairs] of Object.entries(raw.data)) {
-      const arr = new Uint16Array(numBuckets);
+      const arr = new Uint32Array(numBuckets);
       for (const pair of pairs as number[][]) {
         const [bi, cnt] = pair;
         if (bi >= 0 && bi < numBuckets) arr[bi] = cnt;
@@ -38,7 +38,7 @@ function materialize(raw: KillHeatmapResponseRaw): KillHeatmapData {
   } else {
     for (const [sid, dense] of Object.entries(raw.data)) {
       const src = dense as number[];
-      const arr = new Uint16Array(numBuckets);
+      const arr = new Uint32Array(numBuckets);
       for (let i = 0; i < numBuckets; i++) arr[i] = src[i] ?? 0;
       lookup.set(parseInt(sid, 10), arr);
     }
@@ -55,7 +55,7 @@ function materialize(raw: KillHeatmapResponseRaw): KillHeatmapData {
 
 /**
  * Bulk-fetches the per-system kill heatmap dataset for the selected window.
- * Materializes the response into per-system Uint16Array lookups for fast
+ * Materializes the response into per-system Uint32Array lookups for fast
  * indexed access in the render loop. Refetches on (window, space) change.
  */
 export function useKillHeatmap(window: KillHeatmapWindow, space: 'k' | 'w', enabled: boolean) {
